requirements.txt
chainlit
langchain_groq
groq
python-dotenv
-e .